backend/phase1/requirements.txt
django
djangorestframework
django-cors-headers
langchain
langchain_community
langchain_openai
requests
python-dotenv
langchain_anthropic